MODIFY: REFACTOR 2nd , tested and add get address function
