+++ backend/location/routes.py
@@ -1,6 +1,8 @@
 from flask import request,jsonify,Blueprint
 from middleware import *
 from utils import LocationHandle
+from db import db
+import os 
 
 location_bp = Blueprint('location',__name__)
 
@@ -12,12 +14,19 @@ def update_user_location():
     if not data:
         return jsonify({"error": "Request must be JSON"}), 400
     user = request.user
-    location_handle = LocationHandle(data,user)
+    location_handle = LocationHandle(data,user,db)
     latitude,longitude = location_handle.getLocation()
 
     # Set to None if null was explicitly passed, otherwise keep the value
     user.latitude = latitude if latitude is not None else None
     user.longitude = longitude if longitude is not None else None
 
-    return location_handle.saveToDB(latitude,longitude)
+    location_handle.saveToDB(latitude,longitude)
+
+    return jsonify({
+        "message" : "Location updated successfully",
+        "latitude" : user.latitude,
+        "longitude" : user.longitude
+    }),200
     
+
