+++ backend/models.py
@@ -1,6 +1,6 @@
 import uuid
-from app import db
 from werkzeug.security import generate_password_hash, check_password_hash
+from db import db
 
 
 class User(db.Model):
