+++ frontend/index.html
@@ -60,8 +60,10 @@
                 <div class="pdf-upload-area mb-3">
                     <input type="file" id="pdfInput" multiple accept=".pdf" style="display: none;">
                     <button id="uploadPdfBtn" class="btn btn-secondary">Upload PDFs</button>
-                    <button id="shareLocationBtn" style="display: none;" class="btn btn-primary ms-2">Share
-                        Location</button>
+                    <button id="shareLocationBtn" class="btn btn-primary ms-2">Share Location</button>
+                    <div id="address">
+                        <p>My address is: <span id="address-text">Loading...</span></p>
+                    </div>
                     <button id="reasoningBtn" class="btn btn-primary ms-2">Reasoning</button>
                 </div>
                 <div id="uploadedPdfs" class="mt-2 mb-3"></div>
